dom/imptests/webapps/DOMCore/tests/approved/test_Range-collapse.html

<!doctype html>
<title>Range.collapse() and .collapsed tests</title>
<link rel=author title="Aryeh Gregor" href=ayg@aryeh.name>
<div id=log></div>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=common.js></script>
<script>
"use strict";

function testCollapse(rangeEndpoints, toStart) {
	var range;
	if (rangeEndpoints == "detached") {
		range = document.createRange();
		range.detach();
		assert_throws("INVALID_STATE_ERR", function() {
			range.collapse(toStart);
		}, "Detached Range must throw INVALID_STATE_ERR on .collapse()");
		assert_throws("INVALID_STATE_ERR", function() {
			range.collapsed;
		}, "Detached Range must throw INVALID_STATE_ERR on .collapsed access");
		return;
	}

	// Have to account for Ranges involving Documents!
	var ownerDoc = rangeEndpoints[0].nodeType == Node.DOCUMENT_NODE
		? rangeEndpoints[0]
		: rangeEndpoints[0].ownerDocument;
	range = ownerDoc.createRange();
	range.setStart(rangeEndpoints[0], rangeEndpoints[1]);
	range.setEnd(rangeEndpoints[2], rangeEndpoints[3]);

	var expectedContainer = toStart ? range.startContainer : range.endContainer;
	var expectedOffset = toStart ? range.startOffset : range.endOffset;

	assert_equals(range.collapsed, range.startContainer == range.endContainer
		&& range.startOffset == range.endOffset,
		"collapsed must be true if and only if the start and end are equal");

	range.collapse(toStart);

	assert_equals(range.startContainer, expectedContainer,
		"Wrong startContainer");
	assert_equals(range.endContainer, expectedContainer,
		"Wrong endContainer");
	assert_equals(range.startOffset, expectedOffset,
		"Wrong startOffset");
	assert_equals(range.endOffset, expectedOffset,
		"Wrong endOffset");
	assert_true(range.collapsed,
		".collapsed must be set after .collapsed()");
}

var ranges = [
	// Various ranges within the text node children of different paragraphs.
	// (I don't try going out of bounds anywhere here, because these are used
	// for setup, not testing setStart()/setEnd().)
	//
	// Note that the length of the in-document paragraphs is actually 9, due to
	// the newline, but the detached and foreign ones lack the newline, so
	// their length is 8.
	//
	// The format is a bit evil, but it's the only sensible way to output test
	// names.
	"[paras[0].firstChild, 0, paras[0].firstChild, 0]",
	"[paras[0].firstChild, 0, paras[0].firstChild, 1]",
	"[paras[0].firstChild, 2, paras[0].firstChild, 8]",
	"[paras[0].firstChild, 2, paras[0].firstChild, 9]",
	"[paras[1].firstChild, 0, paras[1].firstChild, 0]",
	"[paras[1].firstChild, 0, paras[1].firstChild, 1]",
	"[paras[1].firstChild, 2, paras[1].firstChild, 8]",
	"[paras[1].firstChild, 2, paras[1].firstChild, 9]",
	"[detachedPara1.firstChild, 0, detachedPara1.firstChild, 0]",
	"[detachedPara1.firstChild, 0, detachedPara1.firstChild, 1]",
	"[detachedPara1.firstChild, 2, detachedPara1.firstChild, 8]",
	"[foreignPara1.firstChild, 0, foreignPara1.firstChild, 0]",
	"[foreignPara1.firstChild, 0, foreignPara1.firstChild, 1]",
	"[foreignPara1.firstChild, 2, foreignPara1.firstChild, 8]",
	// Now try testing some elements, not just text nodes.
	"[document.documentElement, 0, document.documentElement, 1]",
	"[document.head, 1, document.head, 1]",
	"[document.body, 4, document.body, 5]",
	"[foreignDoc.documentElement, 0, foreignDoc.documentElement, 1]",
	"[foreignDoc.head, 1, foreignDoc.head, 1]",
	"[foreignDoc.body, 0, foreignDoc.body, 0]",
	"[paras[0], 0, paras[0], 0]",
	"[paras[0], 0, paras[0], 1]",
	"[detachedPara1, 0, detachedPara1, 0]",
	"[detachedPara1, 0, detachedPara1, 1]",
	// Now try some ranges that span elements.
	"[paras[0].firstChild, 0, paras[1].firstChild, 0]",
	"[paras[0].firstChild, 0, paras[1].firstChild, 8]",
	"[paras[0].firstChild, 3, paras[3], 1]",
	// How about something that spans a node and its descendant?
	"[paras[0], 0, paras[0].firstChild, 7]",
	"[testDiv, 2, paras[4], 1]",
	"[testDiv, 1, paras[2].firstChild, 5]",
	"[foreignDoc.documentElement, 1, foreignDoc.body, 0]",
	// Then a few silly things just for good measure.  (Well, document probably
	// isn't entirely silly.)
	"[document, 0, document, 1]",
	"[foreignDoc, 0, foreignDoc, 0]",
	"[xmlDoc, 0, xmlDoc, 0]",
	"[detachedTextNode, 0, detachedTextNode, 8]",
	"[foreignTextNode, 7, foreignTextNode, 7]",
	"[foreignTextNode, 0, foreignTextNode, 8]",
	"[xmlTextNode, 7, xmlTextNode, 7]",
	"[xmlTextNode, 0, xmlTextNode, 8]",
	"[comment, 5, comment, 5]",
	"[foreignComment, 4, foreignComment, 4]",
	"[xmlComment, 2, xmlComment, 6]",
	"[docfrag, 0, docfrag, 0]",
	"[foreignDocfrag, 0, foreignDocfrag, 0]",
	"[xmlDocfrag, 0, xmlDocfrag, 0]",
	// Magic keyword!
	'"detached"'
];

var tests = [];
for (var i = 0; i < ranges.length; i++) {
	tests.push([
		"Range " + i + " " + ranges[i] + ", toStart true",
		eval(ranges[i]),
		true
	]);
	tests.push([
		"Range " + i + " " + ranges[i] + ", toStart false",
		eval(ranges[i]),
		false
	]);
}
generate_tests(testCollapse, tests);

testDiv.style.display = "none";
</script>
